refactor: Rename EmployeeForm to SearchForm
Rename index.js file to completeform.js and create a new index.js.
The reason for this is that there will be a select in the SearchForm
where the user will choose the criteria he'd like to do the search.
After he selects by name, for example, the name field will appear.