feat: Api-server - create saveEmployee function
This function adds a new employee to the data base if
he/she doesnt exist yet, that is, if there is not an
employee in the data base with the same cpf entered by
the user. If that employee exists in the data base,
this function saves the changes made by the user.