feat: SearchForm and Api-server - change the date format to avoid colision with the endpoint
The date query was beeing sent with the format dd/mm/yy. The / in
date query was causing an unexpected behaviour of the api endpoint.
Change date format in SearchForm to dd-mm-yy and in the Api-server
convert that back to dd/mm/yy do match the data base date format.