feat: Context - saveNewEmployeeHandler
Add searchDetails in the state in order to save the last
search criteria and query info of the last search. When user
creates a new employee, add that employee to the searched
Employees results depending on the search criteria and query
of the last search. If a prop in the employee matches the pair
add it to the searchEmployees array.